feat-#98: move sendgrid service to services folder
