+++ src/microservices/mail/sendgrid/sendgrid.module.ts
@@ -1,6 +1,6 @@
 import { Module } from '@nestjs/common';
 import { SendgridController } from './sendgrid.controller';
-import { SendgridService } from './sendgrid.service';
+import { SendgridService } from './services/sendgrid.service';
 
 @Module({
   controllers: [SendgridController],
