+++ src/microservices/mail/sendgrid/services/sendgrid.service.ts
@@ -1,9 +1,9 @@
 import { Inject, Injectable } from '@nestjs/common';
 import { ClientProxy } from '@nestjs/microservices';
 
-import { ClientService } from '../../../utils/client';
-import { SendEmailDto } from './dto';
-import { SuccessResponseModel } from './models';
+import { ClientService } from '../../../../utils/client';
+import { SendEmailDto } from '../dto';
+import { SuccessResponseModel } from '../models';
 
 @Injectable()
 export class SendgridService {
